// OJT_FE/src/types/StatusEnum.ts
export enum ApplyStatus {
  REJECT = "Reject",
  REJECTED = "Rejected",
  APPROVED = "Approved",
  PENDING = "Pending",
}

export enum Status {
  ACTIVE = "Active",
  INACTIVE = "Inactive",
  BANNED = "Banned",

}

export enum Role {
  ADMIN = "Admin",
  STUDENT = "Student",
  HR_STAFF = "HRstaff",
  STAFF = "StudentServicesDepartmentStaff"
}